backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
google-api-core
flask
flask-cors
git rm 
git commit -m "Remove file with API key"
git push origin main
